Finished Project for Loyal Solutions
Spent the 6-8 hours with most of the functionality, and added stuff I thought would make it better afterwards.

My biggest problem was with the multilanguage as I have not worked as much with this. But I was able to find some good sources, to make it possible to create the required functionality.

I had the most fun with the pagination, mostly because I had found a very good resource to make the client-side pagination. Plus I was extremely lucky that the datatable from Vuetify comes with its own pagination :D.